+++ changelog.md
@@ -1,6 +1,7 @@
 # Changelog
 
-- * 3176978 - 2018-11-12: Refactored show object class, cleaned up naming conventions for functions. 
+- * ebe6d3d - 2018-11-12: Updater now waits for show queue to finish before updating. 
+- * 34aa88c - 2018-11-12: Refactored show object class, cleaned up naming conventions for functions. 
 - * 753625d - 2018-11-12: Refactored queue current item property to always represent current running task. 
 - * 8bd128e - 2018-11-11: Release v9.4.30 
 - * 90fdbb1 - 2018-11-11: Pre-Release v9.4.30.dev1